feat(LottoGenerator): 입력된 크기만큼 로또를 생성하는 기능 구현
