+++ docs/README.md
@@ -51,7 +51,7 @@
   - [x] **[예외 처리]** 로또 구입 금액이 1,000원으로 나누어 떨어지지 않으면 예외가 발생한다.
 
 ### 로또 생성 기능
-- [ ] 입력된 크기만큼 로또를 만들 수 있다.
+- [x] 입력된 크기만큼 로또를 만들 수 있다.
 
 ### 보너스 번호 기능
 - [x] 보너스 번호는 1부터 45사이의 숫자이다.
