+++ src/main/java/lotto/domain/LottoStore.java
@@ -1,11 +1,8 @@
 package lotto.domain;
 
-import camp.nextstep.edu.missionutils.Randoms;
-
 import java.util.List;
-import java.util.stream.Stream;
 
-import static lotto.utils.Constants.*;
+import static lotto.utils.Constants.LOTTO_PRICE;
 
 public class LottoStore {
     private static final String LESS_THAN_MINIMUM_PRICE_EXCEPTION_FORMAT = "로또 구입 금액은 최소 %d원 이상이어야 합니다.";
@@ -16,7 +13,7 @@ public class LottoStore {
     public static List<Lotto> buyLotto(int purchaseAmount) {
         validatePurchaseAmount(purchaseAmount);
 
-        return createLottos(purchaseAmount / LOTTO_PRICE);
+        return LottoGenerator.createLottos(purchaseAmount / LOTTO_PRICE);
     }
 
     private static void validatePurchaseAmount(int purchaseAmount) {
@@ -32,19 +29,5 @@ public class LottoStore {
 
     private static boolean isInvalidPurchaseAmount(int purchaseAmount) {
         return purchaseAmount % LOTTO_PRICE != VALID_REMAINING_AMOUNT;
-    }
-
-    private static List<Lotto> createLottos(int size) {
-        return Stream.generate(LottoStore::createLotto)
-                .limit(size)
-                .toList();
-    }
-
-    private static Lotto createLotto() {
-        return new Lotto(pickUniqueNumbers());
-    }
-
-    private static List<Integer> pickUniqueNumbers() {
-        return Randoms.pickUniqueNumbersInRange(MINIMUM_LOTTO_NUMBER, MAXIMUM_LOTTO_NUMBER, LOTTO_NUMBERS_SIZE);
     }
 }
